# generated-sources/src/main/java/com/example/api/StatusApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (5.3.1).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.example.api;

import com.example.model.AccountStatus;
import com.example.model.CustomerStatus;
import com.example.model.InlineResponse500;
import io.swagger.annotations.*;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
@javax.annotation.Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2024-07-04T14:46:36.948950+03:00[Asia/Bahrain]")
@Validated
@Api(value = "Status", description = "the Status API")
public interface StatusApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * GET /accountStatuses/{id} : Retrieve an account status by ID
     * Retrieve an account status by ID
     *
     * @param id  (required)
     * @return Account status details (status code 200)
     *         or Error response (status code 404)
     *         or Error response (status code 500)
     */

    @ApiOperation(value = "Retrieve an account status by ID", nickname = "getAccountStatusById", notes = "Retrieve an account status by ID", response = AccountStatus.class, tags={ "Status", })
    @ApiResponses(value = { 

        @ApiResponse(code = 200, message = "Account status details", response = AccountStatus.class),

        @ApiResponse(code = 404, message = "Error response", response = InlineResponse500.class),

        @ApiResponse(code = 500, message = "Error response", response = InlineResponse500.class) })
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/accountStatuses/{id}",
        produces = { "application/json" }
    )
    default ResponseEntity<AccountStatus> getAccountStatusById(@ApiParam(value = "", required = true) @PathVariable("id") Long id

) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"statusName\" : \"statusName\", \"id\" : 0 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /accountStatuses : Retrieve all account statuses
     * Retrieve all account statuses
     *
     * @return A list of account statuses (status code 200)
     *         or Error response (status code 500)
     */

    @ApiOperation(value = "Retrieve all account statuses", nickname = "getAllAccountStatuses", notes = "Retrieve all account statuses", response = AccountStatus.class, responseContainer = "List", tags={ "Status", })
    @ApiResponses(value = { 

        @ApiResponse(code = 200, message = "A list of account statuses", response = AccountStatus.class, responseContainer = "List"),

        @ApiResponse(code = 500, message = "Error response", response = InlineResponse500.class) })
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/accountStatuses",
        produces = { "application/json" }
    )
    default ResponseEntity<List<AccountStatus>> getAllAccountStatuses() {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"statusName\" : \"statusName\", \"id\" : 0 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /customerStatuses : Retrieve all customer statuses
     * Retrieve all customer statuses
     *
     * @return A list of customer statuses (status code 200)
     *         or Error response (status code 500)
     */

    @ApiOperation(value = "Retrieve all customer statuses", nickname = "getAllCustomerStatuses", notes = "Retrieve all customer statuses", response = CustomerStatus.class, responseContainer = "List", tags={ "Status", })
    @ApiResponses(value = { 

        @ApiResponse(code = 200, message = "A list of customer statuses", response = CustomerStatus.class, responseContainer = "List"),

        @ApiResponse(code = 500, message = "Error response", response = InlineResponse500.class) })
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/customerStatuses",
        produces = { "application/json" }
    )
    default ResponseEntity<List<CustomerStatus>> getAllCustomerStatuses() {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"statusName\" : \"statusName\", \"id\" : 0 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /customerStatuses/{id} : Retrieve a customer status by ID
     * Retrieve a customer status by ID
     *
     * @param id  (required)
     * @return Customer status details (status code 200)
     *         or Error response (status code 404)
     *         or Error response (status code 500)
     */

    @ApiOperation(value = "Retrieve a customer status by ID", nickname = "getCustomerStatusById", notes = "Retrieve a customer status by ID", response = CustomerStatus.class, tags={ "Status", })
    @ApiResponses(value = { 

        @ApiResponse(code = 200, message = "Customer status details", response = CustomerStatus.class),

        @ApiResponse(code = 404, message = "Error response", response = InlineResponse500.class),

        @ApiResponse(code = 500, message = "Error response", response = InlineResponse500.class) })
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/customerStatuses/{id}",
        produces = { "application/json" }
    )
    default ResponseEntity<CustomerStatus> getCustomerStatusById(@ApiParam(value = "", required = true) @PathVariable("id") Long id

) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"statusName\" : \"statusName\", \"id\" : 0 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}