# src/main/java/com/example/accountManager/entity/CustomerStatus.java
package com.example.accountManager.entity;

import jakarta.persistence.*;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.util.Set;

@Data
@NoArgsConstructor
@AllArgsConstructor
@Entity
public class CustomerStatus {
    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Long id;

    @Enumerated(EnumType.STRING)
    private com.example.accountManager.util.CustomerStatus statusName;
    @OneToMany(mappedBy = "status")
    private Set<Customer> customers;
}